# docs/swagger.yaml
definitions:
  main.User:
    properties:
      createdAt:
        type: string
      email:
        type: string
      password:
        type: string
      updatedAt:
        type: string
      userId:
        description: gorm.Model
        type: integer
    type: object
info:
  contact: {}
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Get details of all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.User'
            type: array
      summary: Get details of all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a new user with the input payload
      parameters:
      - description: Create user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
      summary: Create a new user
      tags:
      - users
  /users/{userId}:
    delete:
      consumes:
      - application/json
      description: Delete the user corresponding to the input userId
      parameters:
      - description: ID of the user to be deleted
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
      summary: Delete user identified by the given userId
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update the user corresponding to the input userId
      parameters:
      - description: Update user
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      - description: ID of the user to be updated
        in: path
        name: userId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.User'
      summary: Update user identified by the given userId
      tags:
      - users
swagger: "2.0"